+++ src/freq_data.h
@@ -10,12 +10,12 @@ struct FreqData {
 
     // Note == (12 * Octave + Spectre), where Spectre is in [0, 12)
     std::vector<Color> color; // "Rainbow"-based note color
-    std::vector<double> note; // Note "value" for the frequency
+    std::vector<float> note; // Note "value" for the frequency
     std::vector<int> x; // Horizontal position in the visualization
-    std::vector<double> amp; // Magnitude
+    std::vector<float> amp; // Magnitude
     int minK, maxK; // The range of "meaningful" frequencies
 
     bool ready;
-    std::vector<double> wx;
-    std::vector<double> wy;
+    std::vector<float> wx;
+    std::vector<float> wy;
 };
